chore: rename parseStoreConfig and parseWorldConfig
