+++ packages/world/ts/library/config/index.ts
@@ -1,3 +1,3 @@
-export * from "./parseWorldConfig";
 export * from "./resolveWorldConfig";
 export * from "./types";
+export * from "./worldConfig";
